Module sql_schema_describer::walkers

Functions and types for conveniently traversing and querying a SqlSchema.

Structs

Walker

A generic reference to a schema item. It holds a reference to the schema so it can offer a convenient API based on the Id type.

Type Definitions

ColumnWalker

Traverse a column, that can be in a table or in a view.

DefaultValueWalker

Traverse default value.

EnumVariantWalker

Traverse an enum variant.

EnumWalker

Traverse an enum.

ForeignKeyWalker

Traverse a foreign key.

IndexColumnWalker

Traverse a specific column inside an index.

IndexWalker

Traverse an index.

NamespaceWalker

Traverse a namespace

TableColumnWalker

Traverse a table column.

TableDefaultValueWalker

Traverse a table column’s default value.

TableWalker

Traverse a table.

UserDefinedTypeWalker

Traverse a user-defined type

ViewColumnWalker

Traverse a view column.

ViewDefaultValueWalker

Traverse a view column’s default value.

ViewWalker

Traverse a view