Struct sql_schema_describer::walkers::Walker

pub struct Walker<'a, Id> {
    pub id: Id,
    pub schema: &'a SqlSchema,
}

A generic reference to a schema item. It holds a reference to the schema so it can offer a convenient API based on the Id type.

Fields
id: Id

The identifier.

schema: &'a SqlSchema

The schema for which the identifier is valid.

Implementations

impl<'a, Id> Walker<'a, Id>

pub fn walk<I>(self, other_id: I) -> Walker<'a, I>

Jump to the item identified by other_id.

impl<'a> Walker<'a, ColumnId>

pub fn arity(self) -> ColumnArity

The nullability and arity of the column.

pub fn column_has_enum_default_value(self, enum_name: &str, value: &str) -> bool

Returns whether the column has the enum default value of the given enum type.

pub fn column_type_is_enum(self, enum_name: &str) -> bool

Returns whether the type of the column matches the provided enum name.

pub fn column_type_family(self) -> &'a ColumnTypeFamily

The type family.

pub fn column_type_family_as_enum(self) -> Option<EnumWalker<'a>>

Extract an Enum column type family, or None if the family is something else.

pub fn name(self) -> &'a str

The column name.

pub fn default(self) -> Option<&'a DefaultValue>

The default value for the column.

pub fn column_type(self) -> &'a ColumnType

The full column type.

pub fn column_native_type<T>(self) -> Option<T> where
T: DeserializeOwned,

The column native type.

pub fn is_autoincrement(self) -> bool

Is this column an auto-incrementing integer?

pub fn is_part_of_secondary_index(self) -> bool

Is this column indexed by a secondary index??

pub fn is_part_of_primary_key(self) -> bool

Is this column a part of the table’s primary key?

pub fn is_part_of_foreign_key(self) -> bool

Is this column a part of one of the table’s foreign keys?

pub fn is_same_column(self, other: ColumnWalker<'_>) -> bool

Returns whether two columns are named the same and belong to the same table.

pub fn is_single_primary_key(self) -> bool

Returns whether this column is the primary key. If it is only part of the primary key, this will return false.

pub fn table(self) -> TableWalker<'a>

Traverse to the column’s table.

impl<'a> Walker<'a, ViewId>

pub fn name(self) -> &'a str

The name of the view

pub fn definition(self) -> Option<&'a str>

The SQL definition of the view

impl<'a> Walker<'a, UdtId>

pub fn name(self) -> &'a str

The name of the type

pub fn definition(self) -> Option<&'a str>

The SQL definition of the type

impl<'a> Walker<'a, TableId>

pub fn column(self, column_name: &str) -> Option<ColumnWalker<'a>>

Get a column in the table, by name.

pub fn column_case_insensitive(
self,
column_name: &str
) -> Option<ColumnWalker<'a>>

Get a column in the table, by name.

pub fn columns(self) -> impl ExactSizeIterator<Item = ColumnWalker<'a>>

Traverse the table’s columns.

pub fn foreign_key_count(self) -> usize

The number of foreign key constraints on the table.

pub fn indexes(self) -> impl ExactSizeIterator<Item = IndexWalker<'a>>

Traverse the indexes on the table.

pub fn foreign_keys(self) -> impl ExactSizeIterator<Item = ForeignKeyWalker<'a>>

Traverse the foreign keys on the table.

pub fn referencing_foreign_keys(
self
) -> impl Iterator<Item = ForeignKeyWalker<'a>>

Traverse foreign keys from other tables, referencing current table.

pub fn name(self) -> &'a str

The table name.

pub fn foreign_key_for_column(
self,
column: ColumnId
) -> Option<ForeignKeyWalker<'a>>

Try to traverse a foreign key for a single column.

pub fn namespace(self) -> Option<&'a str>

The namespace the table belongs to, if defined.

pub fn namespace_id(self) -> NamespaceId

The namespace the table belongs to.

pub fn primary_key(self) -> Option<IndexWalker<'a>>

Traverse to the primary key of the table.

pub fn primary_key_columns(
self
) -> Option<impl ExactSizeIterator<Item = IndexColumnWalker<'a>>>

The columns that are part of the primary keys.

pub fn primary_key_columns_count(self) -> usize

How many columns are in the primary key? Returns 0 in the absence of a pk.

impl<'schema> Walker<'a, ForeignKeyId>

pub fn constrained_columns(
self
) -> impl ExactSizeIterator<Item = ColumnWalker<'schema>>

The foreign key columns on the referencing table.

pub fn constraint_name(self) -> Option<&'schema str>

The name of the foreign key constraint.

pub fn on_delete_action(self) -> ForeignKeyAction

The ON DELETE behaviour of the foreign key.

pub fn on_update_action(self) -> ForeignKeyAction

The ON UPDATE behaviour of the foreign key.

pub fn referenced_columns(
self
) -> impl ExactSizeIterator<Item = ColumnWalker<'schema>>

The columns referenced by the foreign key on the referenced table.

pub fn referenced_table_name(self) -> &'schema str

The table the foreign key “points to”.

pub fn referenced_table(self) -> TableWalker<'schema>

The table the foreign key “points to”.

pub fn table(self) -> TableWalker<'schema>

Traverse to the referencing/constrained table.

pub fn is_self_relation(self) -> bool

True if relation is back to the same table.

impl<'a> Walker<'a, IndexColumnId>

pub fn get(self) -> &'a IndexColumn

Get the index column data.

pub fn name(self) -> &'a str

The name of the column.

pub fn length(self) -> Option<u32>

The length limit of the (text) column. Matters on MySQL only.

pub fn sort_order(self) -> Option<SQLSortOrder>

The BTree ordering.

pub fn table(self) -> TableWalker<'a>

The table where the column is located.

pub fn index(self) -> IndexWalker<'a>

The index of the column.

pub fn as_column(self) -> ColumnWalker<'a>

Convert to a normal column walker, losing the possible index arguments.

impl<'a> Walker<'a, IndexId>

pub fn column_names(self) -> impl ExactSizeIterator<Item = &'a str>

The names of the indexed columns.

pub fn columns(self) -> impl ExactSizeIterator<Item = IndexColumnWalker<'a>>

Traverse the indexed columns.

pub fn contains_column(self, column_id: ColumnId) -> bool

True if index contains the given column.

pub fn index_type(self) -> IndexType

The IndexType

pub fn is_primary_key(self) -> bool

Is this index the primary key of the table?

pub fn is_unique(self) -> bool

Is this index a unique constraint? NB: This will return false for the primary key.

pub fn name(self) -> &'a str

The name of the index.

pub fn table(self) -> TableWalker<'a>

Traverse to the table of the index.

impl<'a> Walker<'a, EnumId>

pub fn namespace(self) -> Option<&'a str>

The namespace the enum belongs to, if defined.

pub fn name(self) -> &'a str

The name of the enum. This is a made up name on MySQL.

pub fn values(self) -> &'a [String]

The values of the enum

impl<'a> Walker<'a, NamespaceId>

pub fn name(self) -> &'a str

The namespace name.