Type Definition sql_schema_describer::walkers::ForeignKeyWalker

pub type ForeignKeyWalker<'a> = Walker<'a, ForeignKeyId>;

Traverse a foreign key.

Implementations

impl<'schema> ForeignKeyWalker<'schema>

pub fn constrained_columns(
self
) -> impl ExactSizeIterator<Item = ColumnWalker<'schema>>

The foreign key columns on the referencing table.

pub fn constraint_name(self) -> Option<&'schema str>

The name of the foreign key constraint.

pub fn on_delete_action(self) -> ForeignKeyAction

The ON DELETE behaviour of the foreign key.

pub fn on_update_action(self) -> ForeignKeyAction

The ON UPDATE behaviour of the foreign key.

pub fn referenced_columns(
self
) -> impl ExactSizeIterator<Item = ColumnWalker<'schema>>

The columns referenced by the foreign key on the referenced table.

pub fn referenced_table_name(self) -> &'schema str

The table the foreign key “points to”.

pub fn referenced_table(self) -> TableWalker<'schema>

The table the foreign key “points to”.

pub fn table(self) -> TableWalker<'schema>

Traverse to the referencing/constrained table.

pub fn is_self_relation(self) -> bool

True if relation is back to the same table.