Type Definition sql_schema_describer::walkers::TableWalker

pub type TableWalker<'a> = Walker<'a, TableId>;

Traverse a table.

Implementations

impl<'a> TableWalker<'a>

pub fn column(&self, column_name: &str) -> Option<ColumnWalker<'a>>

Get a column in the table, by name.

pub fn column_case_insensitive(
&self,
column_name: &str
) -> Option<ColumnWalker<'a>>

Get a column in the table, by name.

pub fn columns(self) -> impl ExactSizeIterator<Item = ColumnWalker<'a>>

Traverse the table’s columns.

pub fn foreign_key_count(&self) -> usize

The number of foreign key constraints on the table.

pub fn index_at(&self, index_id: IndexId) -> IndexWalker<'a>

Traverse to an index by id.

pub fn indexes(self) -> impl ExactSizeIterator<Item = IndexWalker<'a>>

Traverse the indexes on the table.

pub fn foreign_keys(self) -> impl Iterator<Item = ForeignKeyWalker<'a>>

Traverse the foreign keys on the table.

pub fn referencing_foreign_keys(
self
) -> impl Iterator<Item = ForeignKeyWalker<'a>>

Traverse foreign keys from other tables, referencing current table.

pub fn name(self) -> &'a str

The table name.

pub fn foreign_key_for_column(
self,
column: &str
) -> Option<ForeignKeyWalker<'a>>

Try to traverse a foreign key for a single column.

pub fn primary_key(self) -> Option<&'a PrimaryKey>

Traverse to the primary key of the table.

pub fn primary_key_columns(
&'a self
) -> Box<dyn ExactSizeIterator<Item = PrimaryKeyColumnWalker<'a>> + 'a>

The columns that are part of the primary keys.

pub fn primary_key_column_names(self) -> Option<Vec<String>>

The names of the columns that are part of the primary key.

pub fn table(self) -> &'a Table

Reference to the underlying Table struct.