Crate sql_schema_describer

Database description. This crate is used heavily in the introspection and migration engines.

Re-exports

pub use self::walkers::*;

Modules

mssql

SQL server schema description.

mysql

MySQL schema description.

postgres

Postgres schema description.

sqlite

SQLite description.

walkers

Functions and types for conveniently traversing and querying a SqlSchema.

Structs

Column

ColumnType

The type of a column.

DefaultValue

DescriberError

Description errors.

EnumId

The identifier for an enum in a SqlSchema.

EnumVariantId

The identifier for an enum variant in a SqlSchema.

ForeignKeyId

The identifier for a foreign key in the schema.

IndexColumn

IndexColumnId

The identifier for a column indexed by a specific index in a SqlSchema.

IndexId

The identifier for an Index in a SqlSchema.

NamespaceId

The identifier for a namespace in the schema.

Procedure

A stored procedure (like, the function inside your database).

SqlMetadata

The return type of get_metadata().

SqlSchema

The result of describing a database schema.

Table

A table found in a schema.

TableColumnId

The identifier for a table column in a SqlSchema.

TableDefaultValueId

The identifier for a table default value in the database.

TableId

The identifier for a table in a SqlSchema.

UdtId

The identifier for a user defined type in the database.

UserDefinedType

A user-defined type. Can map to another type, or be declared as assembly.

View

An SQL view.

ViewColumnId

The identifier for a view column in a SqlSchema.

ViewDefaultValueId

The identifier for a table default value in the database.

ViewId

The identifier for a view in the database.

Enums

ColumnArity

A column’s arity.

ColumnTypeFamily

Enumeration of column type families.

DefaultKind

A DefaultValue

DescriberErrorKind

Variants of DescriberError.

Either

The enum Either with variants Left and Right is a general purpose sum type with two cases.

ForeignKeyAction

Foreign key action types (for ON DELETE|ON UPDATE).

IndexType

The type of an index.

PrismaValue

SQLSortOrder

The sort order of an index.

TableProperties

Traits

SqlSchemaDescriberBackend

A database description connector.

Type Definitions

DescriberResult

The result type.