Module sql_schema_describer::walkers

Functions and types for conveniently traversing and querying a SqlSchema.

Structs

Walker

A generic reference to a schema item. It holds a reference to the schema so it can offer a convenient API based on the Id type.

Type Definitions

ColumnWalker

Traverse column.

EnumWalker

Traverse an enum.

ForeignKeyWalker

Traverse a foreign key.

IndexColumnWalker

Traverse a specific column inside an index.

IndexWalker

Traverse an index.

NamespaceWalker

Traverse a namespace

TableWalker

Traverse a table.

UserDefinedTypeWalker

Traverse a user-defined type

ViewWalker

Traverse a view