Type Definition sql_schema_describer::walkers::ColumnWalker

pub type ColumnWalker<'a> = Walker<'a, ColumnId>;

Traverse column.

Implementations

impl<'a> ColumnWalker<'a>

pub fn arity(self) -> ColumnArity

The nullability and arity of the column.

pub fn column_has_enum_default_value(self, enum_name: &str, value: &str) -> bool

Returns whether the column has the enum default value of the given enum type.

pub fn column_type_is_enum(self, enum_name: &str) -> bool

Returns whether the type of the column matches the provided enum name.

pub fn column_type_family(self) -> &'a ColumnTypeFamily

The type family.

pub fn column_type_family_as_enum(self) -> Option<EnumWalker<'a>>

Extract an Enum column type family, or None if the family is something else.

pub fn name(self) -> &'a str

The column name.

pub fn default(self) -> Option<&'a DefaultValue>

The default value for the column.

pub fn column_type(self) -> &'a ColumnType

The full column type.

pub fn column_native_type<T>(self) -> Option<T> where
T: DeserializeOwned,

The column native type.

pub fn is_autoincrement(self) -> bool

Is this column an auto-incrementing integer?

pub fn is_part_of_secondary_index(self) -> bool

Is this column indexed by a secondary index??

pub fn is_part_of_primary_key(self) -> bool

Is this column a part of the table’s primary key?

pub fn is_part_of_foreign_key(self) -> bool

Is this column a part of one of the table’s foreign keys?

pub fn is_same_column(self, other: ColumnWalker<'_>) -> bool

Returns whether two columns are named the same and belong to the same table.

pub fn is_single_primary_key(self) -> bool

Returns whether this column is the primary key. If it is only part of the primary key, this will return false.

pub fn table(self) -> TableWalker<'a>

Traverse to the column’s table.