Type Definition sql_schema_describer::walkers::TableWalker

pub type TableWalker<'a> = Walker<'a, TableId>;

Traverse a table.

Implementations

impl<'a> TableWalker<'a>

pub fn column(self, column_name: &str) -> Option<TableColumnWalker<'a>>

Get a column in the table, by name.

pub fn column_case_insensitive(
self,
column_name: &str
) -> Option<TableColumnWalker<'a>>

Get a column in the table, by name.

pub fn columns(self) -> impl ExactSizeIterator<Item = TableColumnWalker<'a>>

Traverse the table’s columns.

pub fn foreign_key_count(self) -> usize

The number of foreign key constraints on the table.

pub fn indexes(self) -> impl ExactSizeIterator<Item = IndexWalker<'a>>

Traverse the indexes on the table.

pub fn foreign_keys(self) -> impl ExactSizeIterator<Item = ForeignKeyWalker<'a>>

Traverse the foreign keys on the table.

pub fn referencing_foreign_keys(
self
) -> impl Iterator<Item = ForeignKeyWalker<'a>>

Traverse foreign keys from other tables, referencing current table.

pub fn name(self) -> &'a str

The table name.

pub fn foreign_key_for_column(
self,
column: TableColumnId
) -> Option<ForeignKeyWalker<'a>>

Try to traverse a foreign key for a single column.

pub fn namespace(self) -> Option<&'a str>

The namespace the table belongs to, if defined.

pub fn namespace_id(self) -> NamespaceId

The namespace the table belongs to.

pub fn primary_key(self) -> Option<IndexWalker<'a>>

Traverse to the primary key of the table.

pub fn primary_key_columns(
self
) -> Option<impl ExactSizeIterator<Item = IndexColumnWalker<'a>>>

The columns that are part of the primary keys.

pub fn primary_key_columns_count(self) -> usize

How many columns are in the primary key? Returns 0 in the absence of a pk.

pub fn is_partition(self) -> bool

Is the table a partition table?

pub fn has_subclass(self) -> bool

Does the table have subclasses?