Type Definition sql_schema_describer::walkers::ViewDefaultValueWalker

pub type ViewDefaultValueWalker<'a> = Walker<'a, ViewDefaultValueId>;

Traverse a view column’s default value.

Implementations

impl<'a> ViewDefaultValueWalker<'a>

pub fn coarsen(self) -> DefaultValueWalker<'a>

Coarsen the walker into a generic column default version.

pub fn column(self) -> ViewColumnWalker<'a>

The column where the default value is located.

pub fn as_value(self) -> Option<&'a PrismaValue>

Return a value if a constant.

pub fn as_sequence(self) -> Option<&'a str>

If the value is a squence, return it

pub fn is_value(&self) -> bool

True if a constant value

pub fn is_now(&self) -> bool

True if now()

pub fn is_sequence(&self) -> bool

True if referencing a sequence

pub fn is_db_generated(&self) -> bool

True if value generation is handled in the database

pub fn kind(self) -> &'a DefaultKind

The value kind enumerator

pub fn constraint_name(self) -> Option<&'a str>

The name of the default value constraint.